fail2ban: autodetect log path

+++ update.sh
@@ -14,11 +14,11 @@ source /usr/local/etc/library.sh
 
 # fix automount, reinstall if its old version
 AMFILE=/usr/local/etc/nextcloudpi-config.d/nc-automount.sh
-grep -q inotify-tools $AMFILE || rm $AMFILE
+test -e $AMFILE && { grep -q inotify-tools $AMFILE || rm $AMFILE; }
 
 # fix modsecurity, reinstall if its old verion
 MSFILE=/usr/local/etc/nextcloudpi-config.d/modsecurity.sh
-grep -q "NextCloudPi:" $MSFILE  || rm $MSFILE
+test -e $MSFILE && { grep -q "NextCloudPi:" $MSFILE  || rm $MSFILE; }
 
 # copy all files in bin and etc
 for file in bin/* etc/*; do
